# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: Toolvio Backend API - RBAC & Dynamic Schema System
  version: 1.0.0
  description: |
    Comprehensive API documentation for Toolvio Backend with Role-Based Access Control (RBAC),
    Dynamic Schema Management, Audit Trail, and Multi-tenant Support.
    
    ## 🔐 Authentication
    Most endpoints require JWT authentication. Use the `/api/auth/login` endpoint to get a token,
    then include it in the Authorization header as `Bearer <token>`.
    
    ## 👥 User Roles
    - **Admin**: Full access to all features
    - **Office**: Can manage schemas and data, read audit logs
    - **Technician**: Can read schemas, manage data, read audit logs
    - **Customer**: Can only read data (no schema or audit access)
    
    ## 🏢 Multi-tenant
    All data is isolated by tenant. Users can only access data within their assigned tenant.
    
    Use the "Try it out" button to execute requests against your local server.
servers:
  - url: http://localhost:{port}
    description: Local server
    variables:
      port:
        default: "3000"
paths:
...
  /api/schemas:
    get:
      tags: [Schemas]
      summary: Get all schemas
      description: Retrieve all schemas (requires schema read permission)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: active
          schema: { type: boolean }
          description: Only active schemas when true
      responses:
        '200':
          description: List of schemas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchemaDefinition'
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
    post:
      tags: [Schemas]
      summary: Create new schema
      description: Create a new schema (requires schema write permission + admin/office role)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaDefinitionInput'
            examples:
              basic:
                value:
                  name: product
                  displayName: Product
                  description: Catalog product
                  jsonSchema:
                    type: object
                    properties:
                      name: { type: string }
                      price: { type: number, minimum: 0 }
                    required: [name, price]
      responses:
        '201':
          description: Schema created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDefinition'
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
...
  /api/tenants:
    post:
      tags: [Tenants]
      summary: Create a new tenant
      description: Create a new tenant with the specified configuration
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TenantResponse'
                  message:
                    type: string
        '400':
          description: Invalid request data or tenant already exists
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
    get:
      tags: [Tenants]
      summary: Get all tenants
      description: Retrieve all tenants with pagination, filtering, and sorting
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of items per page
        - in: query
          name: search
          schema:
            type: string
          description: Search term for name, display name, or tenant ID
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
          description: Filter by tenant status
        - in: query
          name: plan
          schema:
            type: string
            enum: [trial, basic, professional, enterprise]
          description: Filter by subscription plan
        - in: query
          name: sortBy
          schema:
            type: string
            default: createdAt
          description: Field to sort by
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Tenants retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tenants:
                        type: array
                        items:
                          $ref: '#/components/schemas/TenantResponse'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: number
                          limit:
                            type: number
                          total:
                            type: number
                          pages:
                            type: number
                  message:
                    type: string
        '401':
          description: Authentication required
        '403':
          description: Insufficient permissions
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


      description: JWT token from /api/auth/login endpoint
...
  schemas:
...
    SchemaDefinition:
      type: object
      properties:
        name: { type: string }
        displayName: { type: string }
        description: { type: string }
        collectionName: { type: string }
        isActive: { type: boolean }
        jsonSchema:
          type: object
    SchemaDefinitionInput:
      type: object
      required: [name, displayName, jsonSchema]
      properties:
        name: { type: string }
        displayName: { type: string }
        description: { type: string }
        jsonSchema:
          type: object
          description: JSON Schema for this entity
...
    TenantCreate:
      type: object
      required:
        - tenantId
        - name
        - displayName
      properties:
        tenantId:
          type: string
          description: Unique tenant identifier
          example: "acme-corp"
        name:
          type: string
          description: Company/organization name
          example: "Acme Corporation"
        displayName:
          type: string
          description: Display name for the tenant
          example: "Acme Corp"
        description:
          type: string
          description: Tenant description
          example: "Leading technology solutions provider"
        contactEmail:
          type: string
          format: email
          description: Primary contact email
          example: "admin@acme.com"
        contactPhone:
          type: string
          description: Primary contact phone
          example: "+1-555-0123"
        settings:
          type: object
          properties:
            maxUsers:
              type: number
              description: Maximum number of users allowed
              example: 100
            maxSchemas:
              type: number
              description: Maximum number of schemas allowed
              example: 50
            maxStorageGB:
              type: number
              description: Maximum storage in GB
              example: 10
        subscriptionPlan:
          type: string
          enum: [trial, basic, professional, enterprise]
          default: trial
          description: Subscription plan
          example: "professional"
...
    TenantResponse:
      type: object
      properties:
        _id:
          type: string
          description: Tenant ID
        tenantId:
          type: string
          description: Unique tenant identifier
        name:
          type: string
          description: Company/organization name
        displayName:
          type: string
          description: Display name for the tenant
        description:
          type: string
          description: Tenant description
        contactEmail:
          type: string
          description: Primary contact email
        contactPhone:
          type: string
          description: Primary contact phone
        settings:
          type: object
          properties:
            maxUsers:
              type: number
              description: Maximum number of users allowed
            maxSchemas:
              type: number
              description: Maximum number of schemas allowed
            maxStorageGB:
              type: number
              description: Maximum storage in GB
            features:
              type: object
              properties:
                auditTrail:
                  type: boolean
                  description: Audit trail feature enabled
                changeStreams:
                  type: boolean
                  description: Change streams feature enabled
                offlineSync:
                  type: boolean
                  description: Offline sync feature enabled
                apiRateLimit:
                  type: boolean
                  description: API rate limiting enabled
        subscriptionPlan:
          type: string
          description: Subscription plan
        isActive:
          type: boolean
          description: Whether tenant is active
        isTrial:
          type: boolean
          description: Whether tenant is in trial period
        trialExpiresAt:
          type: string
          format: date-time
          description: Trial expiration date
        createdAt:
          type: string
          format: date-time
          description: Tenant creation date
        updatedAt:
          type: string
          format: date-time
          description: Last update date
...